add relation between file and jobAnnouncement

+++ backend/src/job-announcement/job.service.ts
@@ -1,20 +1,27 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException, UnauthorizedException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { JobAnnouncement } from 'src/entities/job/jobAnnouncement.entity';
-import { Like, Repository, MoreThanOrEqual, In } from 'typeorm';
+import { Repository } from 'typeorm';
 import { createAnnouncement } from './jobDto/create-announcement.dto';
 import { updateAnnouncement } from './jobDto/update-announcement.dto';
 import { searchAnnouncement } from './jobDto/search-announcement.dto';
 import { Tag } from 'src/entities/job/tag.entity';
 import { User } from 'src/entities/users/user.entity';
+import { CaslAbilityFactory } from 'src/casl/casl-ability.factory';
+import { Action } from 'src/policies/action.enum';
 import { UsersService } from 'src/users/users.service';
+import { FilesService } from 'src/files/files.service';
+import { FileItem } from 'src/entities/files/fileItem.entity';
 
 @Injectable()
 export class JobService {
     constructor(
         @InjectRepository(JobAnnouncement) private readonly repo: Repository<JobAnnouncement>,
         @InjectRepository(Tag) private readonly tagRepo: Repository<Tag>,
-        private UsersService: UsersService
+        @InjectRepository(FileItem) private readonly fileRepo: Repository<FileItem>,
+        private usersService: UsersService,
+        private filesService: FilesService,
+        private readonly caslAbilityFactory: CaslAbilityFactory
     ){}
 
     async index(): Promise<JobAnnouncement[]>{
@@ -63,13 +70,17 @@ export class JobService {
     }
 
     async createAnnouncement(owner : User ,dto: createAnnouncement): Promise<JobAnnouncement>{
-        const {tag, ...announcement} = dto
+        const {tag, pictureId, ...announcement} = dto
         var jobAnnouncement = { ...new JobAnnouncement(), ...announcement };
         jobAnnouncement.tags = [];
-        const user = await this.UsersService.findById(owner.id);
+        const user = await this.usersService.findById(owner.id);
         jobAnnouncement.owner = user;
         var seen = {};
         var tagEntity;
+
+        const picture = await this.filesService.findById(pictureId);
+        if (picture == undefined) throw new NotFoundException();
+        jobAnnouncement.picture = picture
 
         // create relation with tag
         for(var i = 0; i<tag.length; ++i){
@@ -92,10 +103,11 @@ export class JobService {
         return this.repo.save(jobAnnouncement);
     }
 
-    async update(id: number, dto: updateAnnouncement): Promise<JobAnnouncement> {
-        const {tag, ...updateInfo} = dto;
+    async update(owner : User ,id: number, dto: updateAnnouncement): Promise<JobAnnouncement> {
+        const {tag,pictureId, ...updateInfo} = dto;
         var tagArr = [];
         var seen = {};
+        const ability = this.caslAbilityFactory.createForUser(owner);
         if(tag !== undefined){
             for(var i=0; i<tag.length; ++i){
                 await this.tagRepo.findOne({where:{ name: tag[i]}}).then(async (entity)=>{
@@ -116,16 +128,26 @@ export class JobService {
             }
         }
         var announcement, repoAnnouncement
-        await this.findById(id).then(async (entity)=>{
+        await this.repo.findOne(id,{relations:["owner"]}).then(async (entity)=>{
+            if(entity == undefined) throw new NotFoundException();
+            if(!ability.can(Action.Update,entity) && !(owner.id == entity.owner.id)) throw new UnauthorizedException();
             if(tag !== undefined ) entity.tags = tagArr;
+            if(pictureId !== undefined){
+                const picture = await this.fileRepo.findOne(pictureId)
+                if( picture === undefined ) throw new NotFoundException();
+                entity.picture = picture;
+            }
             announcement = {...entity, ...updateInfo}
             repoAnnouncement = await this.repo.save(announcement);
         })
         return repoAnnouncement
     }
 
-    async delete(id: number): Promise<JobAnnouncement>{
-        const jobAnnouncement = await this.findById(id);
+    async delete(owner: User,id: number): Promise<JobAnnouncement>{
+        const ability = this.caslAbilityFactory.createForUser(owner);
+        const jobAnnouncement = await this.repo.findOne(id,{relations:["owner"]});
+        if(jobAnnouncement == undefined) throw new NotFoundException();
+        if(!ability.can(Action.Update,jobAnnouncement) && !(owner.id == jobAnnouncement.owner.id)) throw new UnauthorizedException();
         await this.repo.remove(jobAnnouncement);
         return jobAnnouncement;
     }
